refactor(retry): extract common retry logic to internal/retry package
Moved FixedIntervalBackoff and RetryConfig to shared internal/retry package to eliminate code duplication.

Fix based on code review feedback.
