+++ internal/retry/README.md
@@ -1,0 +1,16 @@
+# retry
+
+Пакет `retry` предоставляет общую логику для повторных попыток выполнения операций с использованием стратегии backoff с фиксированными интервалами.
+
+## Компоненты
+
+- **RetryConfig** - конфигурация для логики повторных попыток
+- **FixedIntervalBackoff** - реализация стратегии backoff с фиксированными интервалами
+- **DefaultRetryConfig()** - функция для получения конфигурации по умолчанию
+
+## Использование
+
+Пакет используется внутри проекта в:
+- `internal/agent` - для повторных попыток HTTP-запросов
+- `internal/repository/postgres` - для повторных попыток операций с базой данных
+
